requirements.txt
alembic===           1.13.1
annotated-types===   0.6.0
anyio  ===           3.7.1
APScheduler===       3.8.0
certifi  ===         2023.11.17
click  ===           8.1.7
colorama   ===       0.4.6
coverage   ===       7.4.0
fastapi   ===        0.105.0
greenlet   ===       3.0.3
h11       ===        0.14.0
httpcore   ===       1.0.2
httptools    ===     0.6.1
httpx      ===       0.26.0
idna      ===        3.6
iniconfig   ===      2.0.0
Mako         ===     1.3.0
MarkupSafe  ===      2.1.3
packaging    ===     23.2
pluggy     ===       1.3.0
pydantic   ===       2.5.3
pydantic_core ===    2.14.6
pytest       ===     7.4.4
python-dotenv  ===   1.0.0
pytz        ===      2023.3.post1
PyYAML      ===      6.0.1
six          ===     1.16.0
sniffio     ===      1.3.0
SQLAlchemy   ===     2.0.23
starlette    ===     0.27.0
typing_extensions=== 4.9.0
tzdata      ===      2023.4
tzlocal     ===      2.1
uvicorn     ===      0.25.0
watchfiles   ===     0.21.0
websockets   ===     12.0
psycopg2       ===   2.9.9

